code structure changed
