+++ src/free.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "libs/Ll.h"
+#include "dataTypes.h"
+
+void FreeVoiceLine(void *ptr);
+
+void FreeChoice(void *ptr);
+
+void FreeScene(void *ptr);
