+++ src/main.c
@@ -7,34 +7,12 @@
 #define CLOGGER_IMPL
 #include "libs/clogger.h"
 //#define JSONLOG
-#define JSON_IMPL
 #include "libs/cJsonParser.h"
-#define ASSLIST_IMPL
 #include "libs/AssList.h"
+#include "dataTypes.h"
+#include "parsers.h"
+#include "gui.h"
 
-#define DEFAULT_W 720
-#define DEFAULT_H 480
-#define SPACING 10*scale
-#define FONT_FILENAME "resources/fonts/OpenSans-Regular.ttf"
-#define TEXTBPSSPEED 10
-
-#define JET (Color){48,48,48, 255}
-void ShowErrorFrontend(const char *errText, const char *elaboration){
-    int screenWidth, screenHeight, screenHalfX, screenHalfY;
-    while(!WindowShouldClose()){
-        ClearBackground(BLACK);
-        screenWidth = GetScreenWidth();
-        screenHeight = GetScreenHeight();
-        screenHalfX = screenWidth/2;
-        screenHalfY = screenHeight/2;
-        float scale = fminf((float)screenHeight/DEFAULT_H, (float)screenWidth/DEFAULT_W);
-        float textSize = scale*20;
-        BeginDrawing();
-            DrawText(errText, screenHalfX - MeasureText(errText, textSize)/2, screenHalfY-3*textSize/2, textSize, RED);
-            DrawText(elaboration, screenHalfX - MeasureText(elaboration, textSize)/2, screenHalfY+3*textSize/2, textSize, RED);
-        EndDrawing();
-    }
-}
 int min(int x, int y){
   return (x < y) ? x : y;
 }
@@ -55,169 +33,6 @@ typedef struct {
     DrawRectangleRoundedLines(btn.rec, 0.5, 5, 2, btn.bdColor); \
     DrawTextEx(gameFont, btn.text, (Vector2){btn.rec.x+5,btn.rec.y+5}, textSize, 1, (!hovered)?btn.bdColor:btn.hoverColorText); \
 }
-
-#define ENTRYCHECK(str){                                                                                    \
-    if(strcmp(jsonEntry, str)!=0){                                                                          \
-        ShowErrorFrontend("[ERROR] Scene format error.", "Game file is broken. Try reinstalling game files.");  \
-        clogger(CLOGGER_ERROR, "scene json format error in %s: got %s\n", str, jsonEntry);                  \
-        exit(1);                                                                                            \
-    }                                                                                                       \
-}
-
-typedef struct {
-    char *name;
-    char *line;
-} VoiceLine;
-
-// function for passing to LlFree(void *ptr)
-void FreeVoiceLine(void *ptr){
-    VoiceLine *vl = (VoiceLine*)ptr;
-    free(vl->name);
-    free(vl->line);
-}
-
-typedef struct {
-    int *id;
-    char *text;
-} Choice;
-
-// function for passing to LlFree(void *ptr)
-void FreeChoice(void *ptr){
-    Choice *vl = (Choice*)ptr;
-    free(vl->id);
-    free(vl->text);
-}
-
-// Next variable: -1 -> scene is last scene
-//                 0 -> end of list
-typedef struct {
-    int number;
-    Llist *characterLl;
-    Texture2D *background;
-    Llist *voiceLineLl;
-    Music *music;
-    Llist *choicesLl;
-    int nextSceneId;
-} Scene;
-
-// function for passing to LlFree(void *ptr)
-void FreeScene(void *ptr){
-    Scene *scene = (Scene*)ptr;
-    LlFree(scene->characterLl);
-    LlFree(scene->voiceLineLl);
-    if(scene->nextSceneId==0){
-        LlFree(scene->choicesLl);
-    }
-    UnloadTexture(*scene->background);
-    UnloadMusicStream(*scene->music);
-}
-
-Llist *ParseChoicesJson(char **gameJson){
-    // TODO: FreeChoice()
-    Llist *choiceLl = malloc(sizeof(Llist));
-    choiceLl->free = FreeChoice;
-    choiceLl->ptr = NULL;
-    choiceLl->size=0;
-    char *jsonEntry=JsonPEAS(gameJson);
-    while(strcmp(jsonEntry, "0")!=0){
-        Choice *choice = malloc(sizeof(*choice));
-        choice->id = malloc(sizeof(int));
-        sscanf(jsonEntry, "%d", choice->id);
-        choice->text = JsonPEAS(gameJson);
-        LlAppend(choiceLl, choice);
-        jsonEntry = JsonPEAS(gameJson);
-    }
-    jsonEntry = JsonPEAS(gameJson);
-    return choiceLl; 
-}
-
-// PEASEs json
-Scene *ParseSceneJson(char **gameJson){
-    Scene *scene = malloc(sizeof(Scene));
-
-    char *jsonEntry = JsonPEAS(gameJson);
-    sscanf(jsonEntry, "%d", &scene->number);
-
-    jsonEntry = JsonPEAS(gameJson);
-    ENTRYCHECK("Background")
-    Texture2D tmpTxt = LoadTexture(JsonPEAS(gameJson));
-    {
-    scene->background = malloc(sizeof(Texture2D));
-    scene->background->id      = tmpTxt.id;
-    scene->background->width   = tmpTxt.width;
-    scene->background->height  = tmpTxt.height;
-    scene->background->mipmaps = tmpTxt.mipmaps;
-    scene->background->format  = tmpTxt.format;
-    }
-
-    jsonEntry = JsonPEAS(gameJson);
-    ENTRYCHECK("Music")
-    Music tmpMusic = LoadMusicStream(JsonPEAS(gameJson));
-    {
-    scene->music = malloc(sizeof(Music));
-    scene->music->stream     = tmpMusic.stream;
-    scene->music->frameCount = tmpMusic.frameCount;
-    scene->music->looping    = tmpMusic.looping;
-    scene->music->ctxType    = tmpMusic.ctxType;
-    scene->music->ctxData    = tmpMusic.ctxData;
-    }
-
-    jsonEntry = JsonPEAS(gameJson);
-    ENTRYCHECK("Characters")
-    scene->characterLl = JsonParseStringArrToLl(gameJson);
-    scene->characterLl->free=free;
-
-    jsonEntry = JsonPEAS(gameJson);
-    ENTRYCHECK("Frases")
-    Llist *lineList = malloc(sizeof(Llist));
-    lineList->ptr  = NULL;
-    lineList->size = 0;
-    lineList->free = &FreeVoiceLine;
-    jsonEntry = JsonPEAS(gameJson);
-    while(strcmp(jsonEntry, "END")!=0){
-        if(strcmp(jsonEntry, "CHOICE")==0){
-            scene->choicesLl = ParseChoicesJson(gameJson);
-            jsonEntry = JsonPEAS(gameJson);
-            scene->nextSceneId=0;
-            continue;
-        }
-        if(strcmp(jsonEntry, "GOTO")==0){
-            jsonEntry = JsonPEAS(gameJson);
-            sscanf(jsonEntry, "%d", &scene->nextSceneId);
-            scene->nextSceneId-=1;
-            jsonEntry = JsonPEAS(gameJson);
-            continue;
-        }
-        VoiceLine *vl = malloc(sizeof(*vl));
-        vl->name = jsonEntry;
-        jsonEntry = JsonPEAS(gameJson);
-        vl->line = jsonEntry;
-        jsonEntry = JsonPEAS(gameJson);
-        LlAppend(lineList, vl);
-    }
-    jsonEntry = JsonPEAS(gameJson);
-    if(strcmp(jsonEntry, "THE END")==0){
-        scene->nextSceneId=-1;
-    }
-    scene->voiceLineLl = lineList;
-    return scene;
-}
-void GuiSlider(Rectangle bounds, float *value, float minValue, float maxValue){
-    float temp = (maxValue - minValue)/2.0f;
-    if (value == NULL) value = &temp;
-    int sliderValue = (int)(((*value - minValue)/(maxValue - minValue))*(bounds.width));
-    Rectangle slider = {bounds.x, bounds.y, 0, bounds.height};
-    slider.width = (float)sliderValue;
-    Vector2 mousePoint = GetMousePosition();
-    if (CheckCollisionPointRec(mousePoint, bounds)&& IsMouseButtonDown(MOUSE_LEFT_BUTTON)){
-        *value = ((maxValue - minValue)*(mousePoint.x - bounds.x))/(float)(bounds.width) + minValue;
-        if (*value > maxValue) *value = maxValue;
-        else if (*value < minValue) *value = minValue;
-    }
-    DrawRectangleRec(bounds, RAYWHITE);
-    DrawRectangleRec(slider, JET);
-}
-
 
 
 //printf("Hello My Froinde!\n"); exit(0);
