+++ src/template.c
